refactor: common-event 모듈 추가
- Event 인터페이스와 EventPublisher 인터페이스 추가
- EventPublisher 구현체 SpringEventPublisher 추가
