+++ event/common-event/build.gradle
@@ -12,8 +12,7 @@ repositories {
 }
 
 dependencies {
-    implementation project(":event:common-event")
-
+    implementation 'org.springframework.boot:spring-boot-starter'
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
