+++ event/common-event/src/main/java/woowa/team4/bff/event/Event.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.event;
+
+public interface Event {
+}
