+++ event/common-event/src/main/java/woowa/team4/bff/publisher/EventPublisher.java
@@ -1,0 +1,8 @@
+package woowa.team4.bff.publisher;
+
+import woowa.team4.bff.event.Event;
+
+public interface EventPublisher {
+
+    void publish(Event event);
+}
