+++ event/common-event/src/main/java/woowa/team4/bff/publisher/SpringEventPublisher.java
@@ -1,7 +1,8 @@
-package woowa.team4.bff.restaurant.event;
+package woowa.team4.bff.publisher;
 
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
+import woowa.team4.bff.event.Event;
 
 @Component
 public class SpringEventPublisher implements EventPublisher {
