+++ event/menu-event/src/main/java/woowa/team4/bff/event/menu/MenuCreateEvent.java
@@ -1,4 +1,6 @@
 package woowa.team4.bff.event.menu;
 
-public record MenuCreateEvent(Long menuId, Long restaurantId, String menuName) {
+import woowa.team4.bff.event.Event;
+
+public record MenuCreateEvent(Long menuId, Long restaurantId, String menuName) implements Event {
 }
