+++ event/restaurant-event/src/main/java/woowa/team4/bff/event/restaurant/RestaurantCreateEvent.java
@@ -1,4 +1,6 @@
 package woowa.team4.bff.event.restaurant;
 
-public record RestaurantCreateEvent(Long restaurantId, String restaurantName) {
+import woowa.team4.bff.event.Event;
+
+public record RestaurantCreateEvent(Long restaurantId, String restaurantName) implements Event {
 }
