--- mono/src/main/java/woowa/team4/bff/restaurant/event/Event.java
@@ -1,4 +1,0 @@
-package woowa.team4.bff.restaurant.event;
-
-public interface Event {
-}
