--- mono/src/main/java/woowa/team4/bff/restaurant/event/EventPublisher.java
@@ -1,6 +1,0 @@
-package woowa.team4.bff.restaurant.event;
-
-public interface EventPublisher {
-
-    void publish(Event event);
-}
